gql_lists_service/pyproject.toml
[tool.poetry]
name = "knowledge-gql-lists"
version = "1.0.0"
description = ""
authors = [
    "Jorgen Ader <jorgen@thorgate.eu>"
]


[tool.poetry.dependencies]
python = "^3.8"
environs = "==7.*"
Flask = "==1.*"
Flask-GraphQL = "==2.*"
Flask-PyMongo = "==2.*"
Flask-JWT-Extended = "==3.*"
graphene-mongo = "==0.2.*"
graphene-federation = "==0.*"


[tool.poetry.dev-dependencies]
black = { version = "^19.10b0", python = "^3.8" }
coverage = "*"
flake8 = "==3.5.*"
isort = "==4.*"
pytest = "==4.*"
pytest-cov = "*"
prospector = "==1.1.6.2"
pycodestyle = "*"
pylint = "<2.3"


[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | docs
  | fabfile.py
  | local.py
  | local_test.py
)
'''

[tool.isort]
skip = 'settings/local.py,settings/local_test.py,fabfile.py,project_light_keepers/wsgi.py,project_light_keepers/asgi.py'
skip_glob="node_modules"
line_length = '88'
atomic = 'true'
multi_line_output = '3'
include_trailing_comma = 'true'
known_django = "django"
known_pytest = "pytest"
known_first_party = "accounts,project_light_keepers"
lines_between_types = '0'
sections = "FUTURE,STDLIB,DJANGO,PYTEST,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
lines_after_imports = '2'
force_alphabetical_sort_within_sections = '1'
